+++ symengine/derivative.cpp
@@ -7,8 +7,8 @@
 #include <symengine/functions.h>
 #include <symengine/integer.h>
 #include <symengine/mul.h>
-#include <symengine/uintpoly.h>
-#include <symengine/uexprpoly.h>
+#include <symengine/polys/uintpoly.h>
+#include <symengine/polys/uexprpoly.h>
 #include <symengine/pow.h>
 #include <symengine/rational.h>
 #include <symengine/sets.h>
