+++ symengine/tests/polynomial/test_mintpoly.cpp
@@ -3,8 +3,8 @@
 #include <iostream>
 
 #include <symengine/mul.h>
-#include <symengine/uintpoly.h>
-#include <symengine/uexprpoly.h>
+#include <symengine/polys/uintpoly.h>
+#include <symengine/polys/uexprpoly.h>
 #include <symengine/polynomial_multivariate.h>
 #include <symengine/pow.h>
 #include <symengine/printer.h>
